S104 split: extract dc_setup.py
